Consolidate platform dependencies
