+++ src/main/kotlin/eu/europa/ec/eudi/sdjwt/Platform.kt
@@ -15,18 +15,38 @@
  */
 package eu.europa.ec.eudi.sdjwt
 
+internal interface Hashes {
+    fun sha256(input: ByteArray): ByteArray
+    fun sha384(input: ByteArray): ByteArray
+    fun sha512(input: ByteArray): ByteArray
+
+    @Suppress("ktlint:standard:function-naming")
+    fun sha3_256(input: ByteArray): ByteArray
+
+    @Suppress("ktlint:standard:function-naming")
+    fun sha3_384(input: ByteArray): ByteArray
+
+    @Suppress("ktlint:standard:function-naming")
+    fun sha3_512(input: ByteArray): ByteArray
+}
+
+internal fun Hashes.digest(hashAlgorithm: HashAlgorithm, input: ByteArray): ByteArray =
+    when (hashAlgorithm) {
+        HashAlgorithm.SHA_256 -> sha256(input)
+        HashAlgorithm.SHA_384 -> sha384(input)
+        HashAlgorithm.SHA_512 -> sha512(input)
+        HashAlgorithm.SHA3_256 -> sha3_256(input)
+        HashAlgorithm.SHA3_384 -> sha3_384(input)
+        HashAlgorithm.SHA3_512 -> sha3_512(input)
+    }
+
+internal interface Random {
+    fun nextBytesCopyTo(bytes: ByteArray)
+}
+
 internal interface Platform {
     val hashes: Hashes
     val random: Random
 }
 
 internal fun platform(): Platform = JvmAndAndroidPlatform
-
-internal object JvmAndAndroidPlatform : Platform {
-
-    override val hashes: Hashes
-        get() = JvmAndAndroidHashes
-
-    override val random: Random
-        get() = JvmAndAndroidSecureRandom
-}
